refactor(adapter-openclaw): extract activity summarization into activity-summary.ts
Fourth cut of the runtime.ts god-file split. Moves the transcript-record →
displayable-ChatChunk layer into a new activity-summary.ts: the
activityChunksFromOpenClawTranscriptRecord / activityChunksFromAssistantMessage
/ activityChunkFromToolResultMessage builders, normalizeToolResultStatus, and
the pure tool-call/shell/URL summarizers + redaction-aware preview helpers
(summarizeOpenClawToolCall/Purpose, summarizeShellCommandPurpose,
summarizeBxContextCommand, summarizeToolNamePurpose, summarizeWebFetchPurpose,
extractUrlForDisplay, cleanShellArgumentForDisplay, basenameForDisplay,
firstLine, previewUnknown, truncateMiddle) + OPENCLAW_ACTIVITY_PREVIEW_CHARS.

The session reader (still in runtime.ts until its own cut) imports back the
one entry point it uses, activityChunksFromOpenClawTranscriptRecord; the rest
stay module-internal.

Also relocates the parseJsonObject leaf (parseJsonValue + isPlainObject, both
already in runtime-utils) into runtime-utils.ts — the moved activity code
depends on it and it was the one JSON leaf missed in the first utils cut;
runtime.ts's other 7 callers now import it from there too.

Pure relocation — bodies byte-for-byte unchanged. runtime.ts 2,600 -> 2,393.
Verified: typecheck, eslint (changed files), adapter suite 115-0 (--isolate),
full suite 5124-0, binary build (stamps reverted), boot smoke (adapter loads,
schedule activates, GET /api/plugins/schedule/ -> 200).
